1.添加定时执行场景待测试

+++ package/usr_app/src/bestomgw/src/dalitek/src/m1_scenario.c
@@ -9,6 +9,8 @@
 #include "m1_protocol.h"
 #include "socket_server.h"
 
+/*函数定义*************************************************************************************/
+static int scenario_alarm_check(scen_alarm_t alarm_time);
 
 /*场景执行*/
 int scenario_exec(char* data, sqlite3* db)
@@ -240,6 +242,7 @@ int scenario_create_handle(payload_t data)
     fprintf(stdout,"scenName:%s\n",scenNameJson->valuestring);
 	/*获取数据包中的alarm信息*/
 	alarmJson = cJSON_GetObjectItem(data.pdu, "alarm");
+
 	/*将alarm信息存入alarm表中*/
 	/*获取table id*/
 	char* sql = "select ID from scen_alarm_table order by ID desc limit 1";
@@ -248,39 +251,44 @@ int scenario_create_handle(payload_t data)
 	/*事物开启*/
 	if(sqlite3_exec(db, "BEGIN", NULL, NULL, &errorMsg)==SQLITE_OK){
         fprintf(stdout,"BEGIN\n");
-		/*获取收到数据包信息*/
-	    hourJson = cJSON_GetObjectItem(alarmJson, "hour");
-	    fprintf(stdout,"hour:%d\n",hourJson->valueint);
-	    minutesJson = cJSON_GetObjectItem(alarmJson, "minutes");
-	    fprintf(stdout,"minutes:%d\n",minutesJson->valueint);
-	    weekJson = cJSON_GetObjectItem(alarmJson, "week");
-	    fprintf(stdout,"week:%s\n",weekJson->valuestring);
-	    statusJson = cJSON_GetObjectItem(alarmJson, "status");
-	    fprintf(stdout,"status:%s\n",statusJson->valuestring);
 	   	/*删除原有表scenario_table中的旧scenario*/
-		sprintf(sql_1,"select ID from scen_alarm_table where SCEN_NAME = \"%s\";",scenNameJson->valuestring);	
-		row_number = sql_row_number(db, sql_1);
-		fprintf(stdout,"row_number:%d\n",row_number);
-		if(row_number > 0){
-			sprintf(sql_1,"delete from scen_alarm_table where SCEN_NAME = \"%s\";",scenNameJson->valuestring);				
-			sqlite3_reset(stmt);
-			sqlite3_prepare_v2(db, sql_1, strlen(sql_1), &stmt, NULL);
-			while(thread_sqlite3_step(&stmt,db) == SQLITE_ROW);
+	   	if(scenNameJson != NULL){
+			sprintf(sql_1,"select ID from scen_alarm_table where SCEN_NAME = \"%s\";",scenNameJson->valuestring);	
+			row_number = sql_row_number(db, sql_1);
+			fprintf(stdout,"row_number:%d\n",row_number);
+			if(row_number > 0){
+				sprintf(sql_1,"delete from scen_alarm_table where SCEN_NAME = \"%s\";",scenNameJson->valuestring);				
+				sqlite3_reset(stmt);
+				sqlite3_prepare_v2(db, sql_1, strlen(sql_1), &stmt, NULL);
+				while(thread_sqlite3_step(&stmt,db) == SQLITE_ROW);
+			}
 		}
+		if(alarmJson != NULL){	
+			/*获取收到数据包信息*/
+		   	hourJson = cJSON_GetObjectItem(alarmJson, "hour");
+		    fprintf(stdout,"hour:%d\n",hourJson->valueint);
+		    minutesJson = cJSON_GetObjectItem(alarmJson, "minutes");
+		    fprintf(stdout,"minutes:%d\n",minutesJson->valueint);
+		    weekJson = cJSON_GetObjectItem(alarmJson, "week");
+		    fprintf(stdout,"week:%s\n",weekJson->valuestring);
+		    statusJson = cJSON_GetObjectItem(alarmJson, "status");
+		    fprintf(stdout,"status:%s\n",statusJson->valuestring);
 		
-	    sql = "insert into scen_alarm_table(ID, SCEN_NAME, HOUR, MINUTES, WEEK, STATUS, TIME) values(?,?,?,?,?,?,?);";
-	    fprintf(stdout,"sql:%s\n",sql);
-	    sqlite3_reset(stmt);
-	    sqlite3_prepare_v2(db, sql, strlen(sql), &stmt, NULL);
-		sqlite3_bind_int(stmt, 1, id);
-		sqlite3_bind_text(stmt, 2, scenNameJson->valuestring, -1, NULL);
-		sqlite3_bind_int(stmt, 3, hourJson->valueint);
-		sqlite3_bind_int(stmt, 4, minutesJson->valueint);
-		sqlite3_bind_text(stmt, 5, weekJson->valuestring, -1, NULL);
-		sqlite3_bind_text(stmt, 6, statusJson->valuestring, -1, NULL);
-		sqlite3_bind_text(stmt, 7, time, -1, NULL);
-		
-		rc = thread_sqlite3_step(&stmt, db); 
+
+		    sql = "insert into scen_alarm_table(ID, SCEN_NAME, HOUR, MINUTES, WEEK, STATUS, TIME) values(?,?,?,?,?,?,?);";
+		    fprintf(stdout,"sql:%s\n",sql);
+		    sqlite3_reset(stmt);
+		    sqlite3_prepare_v2(db, sql, strlen(sql), &stmt, NULL);
+			sqlite3_bind_int(stmt, 1, id);
+			sqlite3_bind_text(stmt, 2, scenNameJson->valuestring, -1, NULL);
+			sqlite3_bind_int(stmt, 3, hourJson->valueint);
+			sqlite3_bind_int(stmt, 4, minutesJson->valueint);
+			sqlite3_bind_text(stmt, 5, weekJson->valuestring, -1, NULL);
+			sqlite3_bind_text(stmt, 6, statusJson->valuestring, -1, NULL);
+			sqlite3_bind_text(stmt, 7, time, -1, NULL);
+			
+			rc = thread_sqlite3_step(&stmt, db); 
+		}
 		
 		/*获取联动表 id*/
 		sql = "select ID from scenario_table order by ID desc limit 1";
@@ -313,7 +321,7 @@ int scenario_create_handle(payload_t data)
 	    		delayJson = cJSON_GetArrayItem(delayArrayJson, j);
 	    		delay += delayJson->valueint;
 	    	}
-	    	fprintf(stdout,"apId:%s, devId:%s, delay:%05d\n",apIdJson->valuestring, devIdJson->valuestring, delayJson->valueint);
+	    	fprintf(stdout,"apId:%s, devId:%s\n",apIdJson->valuestring, devIdJson->valuestring);
 	    	paramArrayJson = cJSON_GetObjectItem(devJson, "param");
 	    	number2 = cJSON_GetArraySize(paramArrayJson);
 	    	for(j = 0; j < number2; j++){
@@ -797,4 +805,76 @@ int app_req_scenario_name(int clientFd, int sn)
 
 }
 
+/*定时执行场景检查*/
+ void scenario_alarm_select(void)
+ {
+ 	printf("scenario_alarm_select\n");
+ 	int rc;
+ 	char * sql = NULL;
+ 	scen_alarm_t alarm;
+ 	while(1){
+ 		sqlite3_stmt* stmt = NULL;
+ 		sqlite3* db = NULL;
+	 	rc = sqlite3_open("dev_info.db", &db);  
+	     if( rc ){  
+	         fprintf(stderr, "Can't open database: %s\n", sqlite3_errmsg(db));  
+	         return M1_PROTOCOL_FAILED;  
+	     }else{  
+	         fprintf(stderr, "Opened database successfully\n");  
+	     }
+	     char* sql = "select SCEN_NAME, HOUR, MINUTES, WEEK, STATUS from scen_alarm_table;";
+	     sqlite3_prepare_v2(db, sql, strlen(sql), &stmt, NULL);
+	     while(thread_sqlite3_step(&stmt, db) == SQLITE_ROW){
+	     	alarm.scen_name = sqlite3_column_text(stmt, 0);
+	     	alarm.hour = sqlite3_column_int(stmt, 1);
+	     	alarm.minutes = sqlite3_column_int(stmt, 2);
+	     	alarm.week = sqlite3_column_text(stmt, 3);
+	     	alarm.status = sqlite3_column_text(stmt, 4);
+	     	rc = scenario_alarm_check(alarm);
+	     	if(rc)
+	     		scenario_exec(alarm.scen_name, db);
+	     }
+	     sqlite3_finalize(stmt);
+	     sqlite3_close(db);
 
+	     sleep(60);
+ 	}
+ }
+
+ static int scenario_alarm_check(scen_alarm_t alarm_time)
+ {
+ 	int on_time_flag = 0;
+ 	char* week = NULL;
+ 	struct tm nowTime;
+     struct timespec time;
+     clock_gettime(CLOCK_REALTIME, &time);  //获取相对于1970到现在的秒数
+     localtime_r(&time.tv_sec, &nowTime);
+     if(strcmp(alarm_time.status,"on") == 0){
+     	switch(nowTime.tm_wday){
+     		case 0: week = "sunday";break;
+     		case 1: week = "monday";break;
+     		case 2: week = "tuesday";break;
+     		case 3: week = "wednessday";break;
+     		case 4: week = "thursday";break;
+     		case 5: week = "friday";break;
+     		case 6: week = "saturday";break;	
+     	}
+     	/*周一到周日*/
+     	if((strcmp(alarm_time.week, week) == 0) || (strcmp(alarm_time.week, "all") == 0)){
+     		on_time_flag = 1;
+     	}else if((nowTime.tm_wday != 6) && (nowTime.tm_wday != 0) && (strcmp(alarm_time.week, "workDay") == 0)){               //周一到周五
+     		on_time_flag = 1;
+     	}else{
+     		on_time_flag = 0;
+     	}
+     	/*小时、分钟*/
+     	if((alarm_time.hour == nowTime.tm_hour) && (alarm_time.minutes == nowTime.tm_min) && on_time_flag){
+     		on_time_flag = 1;
+     	}else{
+     		on_time_flag = 0;
+     	}
+     	
+     }
+     //on_time_flag = 1;
+     return on_time_flag;
+ }
